updating reviewer questions to add loading/skeleton state txstate-etc/reqquest-txstate#376
