+++ demos/src/rc/definitions/prompts/applicationManagement.prompts.ts
@@ -4,6 +4,13 @@ import { MutationMessageType } from '@txstate-mws/graphql-server'
 import { fileHandler } from 'fastify-txstate'
 import { OptOutData, OptOutSchema } from '../models/optOut.models.js'
 
+const waitFiveSeconds = () => {
+    return new Promise((resolve) => {
+        setTimeout(resolve, 5000);
+    });
+};
+
+
 export const application_management_opt_out_prompt: PromptDefinition<OptOutData> = {
   key: 'application_management_opt_out_prompt',
   title: 'Application management',
@@ -39,6 +46,13 @@ export const assess_technical_troubleshooting_prompt: PromptDefinition<AssessTec
     if (data.demonstrateTechincalTroubleshooting == null) messages.push({ type: MutationMessageType.error, message: 'This field is required', arg: '.demonstrateTechincalTroubleshooting' })
 
     return messages
+  },
+  preload: async (data, config) => {
+    await waitFiveSeconds() // simulate a long-running preload operation
+    return {
+      demonstrateTechincalTroubleshooting: true,
+      complexity: 1
+    }
   }
 }
