+++ ui/src/routes/requests/[id]/approve/[programKey]/+page.svelte
@@ -1,25 +1,21 @@
 <script lang="ts">
-  import { BadgeNumber, FieldCheckbox, FieldTextArea, FormInlineNotification, Panel, PanelDialog, PanelFormDialog } from '@txstate-mws/carbon-svelte'
+  import { BadgeNumber, FieldTextArea, PanelDialog, PanelFormDialog } from '@txstate-mws/carbon-svelte'
   import { toasts } from '@txstate-mws/svelte-components'
-  import { Form } from '@txstate-mws/svelte-forms'
-  import { Button, InlineNotification, Select, SelectItem, Tooltip } from 'carbon-components-svelte'
+  import ReviewerQuestions from '$internal/components/ReviewerQuestions.svelte'
+  import { Button, InlineNotification, Select, SelectItem } from 'carbon-components-svelte'
   import DocumentExport from 'carbon-icons-svelte/lib/DocumentExport.svelte'
   import Edit from 'carbon-icons-svelte/lib/Edit.svelte'
-  import MachineLearning from 'carbon-icons-svelte/lib/MachineLearning.svelte'
   import Pen from 'carbon-icons-svelte/lib/Pen.svelte'
   import TrashCan from 'carbon-icons-svelte/lib/TrashCan.svelte'
   import View from 'carbon-icons-svelte/lib/View.svelte'
-  import WarningAltFilled from 'carbon-icons-svelte/lib/WarningAltFilled.svelte'
-  import WarningFilled from 'carbon-icons-svelte/lib/WarningFilled.svelte'
   import { invalidateAll } from '$app/navigation'
-  import { api, RenderDisplayComponent, applicantRequirementTypes, reviewerRequirementTypes, getApplicationStatusInfo } from '$internal'
-  import { CommentCard, enumPromptVisibility, enumRequirementStatus, enumRequirementType, InfoCard } from '$lib'
+  import { api, getApplicationStatusInfo } from '$internal'
+  import { CommentCard, enumPromptVisibility, enumRequirementStatus, enumRequirementType, InfoCard, PromptIndicators } from '$lib'
   import type { PageData } from './$types'
   import { uiRegistry } from '../../../../../local'
-  import { isInlineReviewerEditPrompt } from '../../../../../internal'
   import ApproveLayout from '../ApproveLayout.svelte'
   import { Loading } from "carbon-components-svelte";
-  import { load } from '../+page';
+
 
   /**
    * This page is the primary reviewer screen, it will show all the prompt
@@ -98,11 +94,6 @@
   }
 
   type ApplicationRequirement = (typeof appRequest)['applications'][number]['requirements'][number]
-  const PromptIndicators = {
-    AUTOMATION: 1,
-    WARNING: 2,
-    DISQUALIFYING: 3
-  } as const
 
   // here we split the screen into sections based on the requirement type, but we
   // keep everything in order, so if a reviewer requirement appears between two
@@ -188,59 +179,6 @@
     }
   }
 
-  type PromptExtraData = Awaited<ReturnType<typeof api.getPromptData>>
-  type Prompt = PageData['appRequest']['applications'][0]['requirements'][0]['prompts'][0]
-  type PromptWithExtra = Prompt & PromptExtraData
-  let promptBeingEdited: PromptWithExtra | undefined = undefined
-  let showPromptDialog = false
-  let fetchingEditPrompt = false
-  function editPrompt (prompt: Prompt) {
-    return async () => {
-      if (fetchingEditPrompt) return
-      fetchingEditPrompt = true
-      try {
-        const extra = await api.getPromptData(appRequest.id, prompt.id)
-        promptBeingEdited = { ...prompt, ...extra }
-        showPromptDialog = true
-      } finally {
-        fetchingEditPrompt = false
-      }
-    }
-  }
-  function hideEditModalPromptOnLoading() { // keep the dialog in the DOM so onSave fires, but remove from view
-    if (!showPromptDialog) return
-    const editModalDialog = document.querySelector('.tcbs-dialog')
-    editModalDialog?.classList.add('invisible')   
-  }
-
-  function closePromptDialog () {
-    fetchingEditPrompt = false
-    showPromptDialog = false
-    promptBeingEdited = undefined
-  }
-
-  function onPromptSubmit (id: string) {
-    return async (data: any) => {   
-      loading = true      
-      hideEditModalPromptOnLoading()  
-      const response = await api.updatePrompt(id, data, false)
-      return response
-    }
-  }
-
-  function onPromptValidate (prompt: any) {
-    return async (data: any) => {
-      const response = await api.updatePrompt(prompt.id, data, true)
-      return response.messages
-    }
-  }
-
-  async function onPromptSaved (data: any) {        
-    await invalidateAll()     
-    closePromptDialog ()
-    loading = false
-  }
-
   let appAction: '' | 'advanceWorkflow' | 'reverseWorkflow' = ''
   async function onAppAction () {
     if (appAction === 'advanceWorkflow') {
@@ -320,74 +258,12 @@
           noborder
           />
         {:else}
-          <InlineNotification kind="info" lowContrast hideCloseButton=true title="No application notes." subtitle="Add a note to see it here."></InlineNotification>
+          <InlineNotification kind="info" lowContrast hideCloseButton title="No application notes." subtitle="Add a note to see it here."></InlineNotification>
         {/if}
       </div>
     </InfoCard>
   </svelte:fragment>
-  {#each sections as section (section.title)}
-    <Panel title={section.title} expandable expanded>
-      {#if section.requirements.some(r => r.prompts.length > 0)}
-        <dl class="prompts">
-          {#each section.requirements as requirement (requirement.id)}
-            {#each requirement.prompts as prompt (prompt.id)}
-              {@const def = uiRegistry.getPrompt(prompt.key)}
-              {@const isReviewerQuestion = reviewerRequirementTypes.has(requirement.type) && !def?.automation}
-              {@const isAutomation = !!def?.automation}
-              {@const editMode = isInlineReviewerEditPrompt(def, requirement, prompt)}
-              {@const small = editMode && def.formMode !== 'full' ? def.formMode !== 'large' : def!.displayMode !== 'large'}
-              {@const large = editMode && def.formMode !== 'full' ? def.formMode === 'large' : def!.displayMode === 'large'}
-              {@const dtid = `dt-title-${prompt.id}`}
-              <dt class:small class:large class:isReviewerQuestion class:bg-tagyellow-200={isAutomation}>
-                {#if promptIndicator[prompt.key]?.indicator}
-                  <div class="indicator-tooltip">
-                    <Tooltip align="start" direction="bottom">
-                      <svelte:fragment slot="icon">
-                        {#if promptIndicator[prompt.key]?.indicator === PromptIndicators.AUTOMATION}
-                          <MachineLearning size={20} />
-                        {:else if promptIndicator[prompt.key]?.indicator === PromptIndicators.WARNING}
-                          <WarningAltFilled size={20} class="warning-icon" />
-                        {:else if promptIndicator[prompt.key]?.indicator === PromptIndicators.DISQUALIFYING}
-                          <WarningFilled size={20} class="disqualifying-icon" />
-                        {/if}
-                      </svelte:fragment>
-                      {promptIndicator[prompt.key]?.reason}
-                    </Tooltip>
-                  </div>
-                {/if}
-                <div id={dtid}>
-                  {prompt.title}
-                </div>
-              </dt>
-              <dd class="flow" class:small class:large class:isReviewerQuestion class:bg-tagyellow-200={isAutomation} role={editMode ? 'group' : undefined} aria-labelledby={dtid}>
-                {#if editMode}
-                  <Form preload={prompt.preloadData} submit={onPromptSubmit(prompt.id)} validate={onPromptValidate(prompt)} autoSave on:autosaved={onPromptSaved} let:data let:messages>
-                    <svelte:component this={def.formComponent} {data} appRequestData={appRequest.data} prestageData={{latest: prompt.prestageData, current: appRequest.data[prompt.key]?.__prestage}} fetched={prompt.fetchedData} configData={prompt.configurationData} gatheredConfigData={prompt.gatheredConfigData} />
-                    {#each messages as message (message.message, message.type)}
-                      <FormInlineNotification {message} />
-                    {/each}
-                  </Form>
-                {:else}
-                  <RenderDisplayComponent {def} appRequestId={appRequest.id} appData={appRequest.data} prompt={prompt} prestageData={{latest: prompt.prestageData, current: appRequest.data[prompt.key]?.__prestage}} configData={prompt.configurationData} gatheredConfigData={prompt.gatheredConfigData} showMoot />
-                  {#if prompt.actions.update}
-                    {#if prompt.invalidated && !applicantRequirementTypes.has(requirement.type)}
-                      <Button kind="primary" size="field" class="prompt-edit" on:click={editPrompt(prompt)}>Review correction</Button>
-                    {:else}
-                      <Button kind="ghost" size="field" icon={Edit} iconDescription="Edit Prompt" class="prompt-edit" on:click={editPrompt(prompt)} />
-                    {/if}
-                  {/if}
-                {/if}
-              </dd>
-            {/each}
-          {/each}
-        </dl>
-      {:else if section.requirements[0]?.workflowStage && section.requirements[0]?.workflowStage.key === application.workflowStage?.key}
-        No questions need to be answered in this section. You may advance to the next step.
-      {:else}
-        No questions in this section.
-      {/if}
-    </Panel>
-  {/each}
+  <ReviewerQuestions {sections} {appRequest} {application} {promptIndicator} />
   <div class="app-actions [ flex items-end ]">
     {#if application.actions.advanceWorkflow || application.actions.reverseWorkflow}
       <Select bind:selected={appAction} labelText="Next step" size="sm">
@@ -405,26 +281,7 @@
   </div>
 </ApproveLayout>
 
-{#if showPromptDialog && promptBeingEdited}
-  <PanelFormDialog
-    title={promptBeingEdited.invalidated ? `Review correction "${promptBeingEdited.title}"` : 'Edit Prompt'}
-    bind:open={showPromptDialog}
-    on:cancel={closePromptDialog}
-    submit={onPromptSubmit(promptBeingEdited.id)}
-    validate={onPromptValidate(promptBeingEdited)}
-    on:saved={onPromptSaved}
-    disableSaveUntilChanged={true}
-    centered
-    preload={promptBeingEdited.preloadData}
-    let:data
-  >
-    {@const def = uiRegistry.getPrompt(promptBeingEdited.key)}
-    <div class='font-medium text-center mt-2'>
-      <p class="text-xl font-medium ">{promptBeingEdited.title}</p>
-    </div>
-    <svelte:component this={def!.formComponent} appRequestId={appRequest.id} {data} appRequestData={promptBeingEdited.data} prestageData={{latest: promptBeingEdited.prestageData, current: appRequest.data[promptBeingEdited.key]?.__prestage}} fetched={promptBeingEdited.fetchedData} configData={promptBeingEdited.configurationData} gatheredConfigData={promptBeingEdited.gatheredConfigData} />
-  </PanelFormDialog>
-{/if}
+
 
 {#if showUpdateNoteDialog && noteBeingEdited}
   <PanelFormDialog
@@ -499,55 +356,6 @@
 </PanelDialog>
 
 <style>
-  dl.prompts {
-    display: grid;
-    grid-template-columns: auto 1fr;
-    align-items: stretch;
-    gap: 0;
-    margin: -16px;
-  }
-  .prompts dt, .prompts dd {
-    position: relative;
-    border-bottom: 1px solid var(--cds-border-subtle);
-    padding: 1rem 15px;
-  }
-  .prompts dt.small {
-    padding-right: 15px;
-  }
-  .prompts dt.large, .prompts dd.large {
-    grid-column: span 2;
-  }
-  .prompts dt.isReviewerQuestion, .prompts dd.isReviewerQuestion {
-    background-color: var(--cds-tag-background-cyan);
-  }
-  .prompts dt {
-    display: flex;
-    align-items: center;
-    gap: 4px;
-    padding-left: 32px;
-  }
-  .prompts dd :global(.bx--btn.prompt-edit) {
-    position: absolute;
-    top: 0;
-    right: 0;
-  }
-  .prompts dt .indicator-tooltip {
-    display: inline-block;
-    margin-left: -24px;
-  }
-  .prompts dt .indicator-tooltip :global(> div) {
-    line-height: 0;
-  }
-  .prompts dt .indicator-tooltip :global(> div .bx--tooltip--shown) {
-    line-height: 1.2;
-  }
-  .prompts dt :global(.warning-icon) {
-    fill: var(--cds-inverse-support-03, #f1c21b);
-  }
-  .prompts dt :global(.disqualifying-icon) {
-    fill: var(--cds-support-01, #da1e28);
-  }
-
   .active-note {
     padding: 4px 8px;
   }
